made changes on the authentication flow
